fix(hw2): 0203 use 32 bits int except for final result (days)

+++ hw2/hw0203.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 #include <stdint.h>
 
-int64_t validate_input(int64_t yy, int64_t mm, int64_t dd, int64_t h, int64_t m);
-int64_t calc_total_days(int64_t y1, int64_t m1, int64_t d1, int64_t y2, int64_t m2, int64_t d2);
-int8_t calc_weekday(int64_t y, int64_t m, int64_t d);
-int64_t calc_single(int64_t h1, int64_t m1, int64_t h2, int64_t m2);
+int32_t validate_input(int32_t yy, int32_t mm, int32_t dd, int32_t h, int32_t m);
+int32_t calc_total_days(int32_t y1, int32_t m1, int32_t d1, int32_t y2, int32_t m2, int32_t d2);
+int8_t calc_weekday(int32_t y, int32_t m, int32_t d);
+int32_t calc_single(int32_t h1, int32_t m1, int32_t h2, int32_t m2);
 
-int64_t main() {
-    int64_t b_yy, b_mm, b_dd, b_h, b_m, e_yy, e_mm, e_dd, e_h, e_m;
+int32_t main() {
+    int32_t b_yy, b_mm, b_dd, b_h, b_m, e_yy, e_mm, e_dd, e_h, e_m;
 
     printf("From: ");
     // Throw an ERROR if the input is not accepted
-    if (scanf("%ld/%ld/%ld %ld:%ld", &b_yy, &b_mm, &b_dd, &b_h, &b_m) != 5) {
+    if (scanf("%d/%d/%d %d:%d", &b_yy, &b_mm, &b_dd, &b_h, &b_m) != 5) {
         printf("Invalid Input: Only Integers Allowed.\n");
         return 1;
     }
 
     printf("To: ");
     // Throw an ERROR if the input is not accepted
-    if (scanf("%ld/%ld/%ld %ld:%ld", &e_yy, &e_mm, &e_dd, &e_h, &e_m) != 5) {
+    if (scanf("%d/%d/%d %d:%d", &e_yy, &e_mm, &e_dd, &e_h, &e_m) != 5) {
         printf("Invalid Input: Only Integers Allowed.\n");
         return 1;
     }
@@ -33,7 +33,7 @@ int64_t main() {
     }
 
     int64_t result = 0;
-    int64_t total_days = calc_total_days(b_yy, b_mm, b_dd, e_yy, e_mm, e_dd);
+    int32_t total_days = calc_total_days(b_yy, b_mm, b_dd, e_yy, e_mm, e_dd);
     // printf("[DEBUG] total_days: %ld\n", total_days);
 
     int8_t weekday = calc_weekday(b_yy, b_mm, b_dd);
@@ -54,7 +54,7 @@ int64_t main() {
     }
     else {
         if (weekday != 0 && weekday != 6) result += calc_single(b_h, b_m, 24, 0);
-        for (int64_t i = 1; i < total_days; i++) {
+        for (int32_t i = 1; i < total_days; i++) {
             weekday = (weekday + 1) % 7;
             if (weekday != 0 && weekday != 6) result += calc_single(0, 0, 24, 0);
         }
@@ -69,14 +69,14 @@ int64_t main() {
 }
 
 // check if the input year is leap year. @return yes(1) or no(0)
-int64_t is_leap(int64_t yr) {
+int32_t is_leap(int32_t yr) {
     if (yr % 4 == 0 && yr % 100 != 0) return 1;
     if (yr % 400 == 0) return 1;
     return 0;
 }
 
 // validate the input. @return valid(0) or invalid(1)
-int64_t validate_input(int64_t yy, int64_t mm, int64_t dd, int64_t h, int64_t m) {
+int32_t validate_input(int32_t yy, int32_t mm, int32_t dd, int32_t h, int32_t m) {
     if (yy < 0 || mm < 0 || mm > 12 || dd < 0 || dd > 31 || h < 0 || h > 23 || m < 0 || m > 59) return 1;
     if ((mm == 4 || mm == 6 || mm == 9 || mm == 11) && dd > 30) return 1;
     if (mm == 2 && dd > is_leap(yy) + 28) return 1;
@@ -85,31 +85,31 @@ int64_t validate_input(int64_t yy, int64_t mm, int64_t dd, int64_t h, int64_t m)
 
 // calculate the first day's weekday (Thanks to the Internet) 
 // @return weekday: Monday = 1, Tuesday = 2, Wednesday = 3, Thursday = 4, Friday = 5, Saturday = 6, Sunday = 0
-int8_t calc_weekday(int64_t y, int64_t m, int64_t d) {
+int8_t calc_weekday(int32_t y, int32_t m, int32_t d) {
     return (d += m < 3 ? y-- : y - 2, 23 * m / 9 + d + 4 + y / 4 - y / 100 + y / 400) % 7;
     // m = (m == 1 || m == 2) ? m + 10 : m - 2;
-    // int64_t c = y / 100 - 1;
+    // int32_t c = y / 100 - 1;
     // y -= (m == 1 || m == 2) ? 1 : 0;
     // int8_t weekday = (((y % 100) * 5 / 4) + m + d - 2 * (c % 4) + 7) % 7;
     // return weekday;
 }
 
 // calculate the number of days between two dates in one specific year. @return days
-int64_t calc_days_in_year(int64_t yr, int64_t m1, int64_t d1, int64_t m2, int64_t d2) {
-    int64_t month[12] = { 31, 28 + is_leap(yr), 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
+int32_t calc_days_in_year(int32_t yr, int32_t m1, int32_t d1, int32_t m2, int32_t d2) {
+    int32_t month[12] = { 31, 28 + is_leap(yr), 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
 
-    int64_t days = 0;
-    for (int64_t i = m1; i < m2; i++) days += month[i];
+    int32_t days = 0;
+    for (int32_t i = m1; i < m2; i++) days += month[i];
     days += d2 - d1;
     return days;
 }
 
 // calculate the number of days between two dates. @return days
-int64_t calc_total_days(int64_t y1, int64_t m1, int64_t d1, int64_t y2, int64_t m2, int64_t d2) {
-    int64_t days = 0;
+int32_t calc_total_days(int32_t y1, int32_t m1, int32_t d1, int32_t y2, int32_t m2, int32_t d2) {
+    int32_t days = 0;
 
     // years between y1 and y2
-    for (int64_t i = y1 + 1; i < y2; i++) days += is_leap(i) + 365;
+    for (int32_t i = y1 + 1; i < y2; i++) days += is_leap(i) + 365;
 
     if (y1 == y2) {
         days += calc_days_in_year(y1, m1, d1, m2, d2);
@@ -123,12 +123,12 @@ int64_t calc_total_days(int64_t y1, int64_t m1, int64_t d1, int64_t y2, int64_t 
 }
 
 // calculate the working minutes of a single day @return minutes
-int64_t calc_single(int64_t h1, int64_t m1, int64_t h2, int64_t m2) {
-    int64_t begin = h1 * 60 + m1, end = h2 * 60 + m2;
+int32_t calc_single(int32_t h1, int32_t m1, int32_t h2, int32_t m2) {
+    int32_t begin = h1 * 60 + m1, end = h2 * 60 + m2;
     if (begin < 9 * 60) begin = 9 * 60;
     if (end > 18 * 60 + 30) end = 18 * 60 + 30;
 
-    int64_t diff = end - begin;
+    int32_t diff = end - begin;
 
     if (begin < 12 * 60 && end > 13 * 60 + 30) diff -= 90;
     else if (begin < 12 * 60) diff -= (end - 12 * 60);
